Config file handling changes - name changed back to .pastaELN.json - extractors and project are moved into project group - current project group is attribute of backend - sqlite:update: if no tags in data: skip that part - test api key has two failure modes

+++ pasta_eln/sqlite.py
@@ -345,28 +345,30 @@ class SqlLiteDB:
     cursor = self.connection.cursor()
 
     # tags and qrCodes
-    tagsNew= set(dataNew.pop('tags'))
-    self.cursor.execute(f"SELECT tag FROM tags WHERE id == '{docID}'")
-    tagsOld= {i[0] for i in self.cursor.fetchall()}
-    if tagsOld.difference(tagsNew):
-      cmd = f"DELETE FROM tags WHERE id == '{docID}' and tag == ?"
-      self.cursor.executemany(cmd, [(i,) for i in tagsOld.difference(tagsNew)])
-      changesDict['tags'] = ','.join(tagsOld)
-    if tagsNew.difference(tagsOld):
-      change = tagsNew.difference(tagsOld)
-      self.cursor.executemany("INSERT INTO tags VALUES (?, ?);", zip([docID]*len(change), change))
-      changesDict['tags'] = ','.join(tagsOld)
-    qrCodesNew= set(dataNew.pop('qrCodes', []))
-    self.cursor.execute(f"SELECT qrCode FROM qrCodes WHERE id == '{docID}'")
-    qrCodesOld= {i[0] for i in self.cursor.fetchall()}
-    if qrCodesOld.difference(qrCodesNew):
-      cmd = f"DELETE FROM qrCodes WHERE id == '{docID}' and qrCode == ?"
-      self.cursor.executemany(cmd, [(i,) for i in qrCodesOld.difference(qrCodesNew)])
-      changesDict['qrCodes'] = ','.join(qrCodesOld)
-    if qrCodesNew.difference(qrCodesOld):
-      change = qrCodesNew.difference(qrCodesOld)
-      self.cursor.executemany("INSERT INTO qrCodes VALUES (?, ?);", zip([docID]*len(change), change))
-      changesDict['qrCodes'] = ','.join(qrCodesOld)
+    if 'tags' in dataNew:
+      tagsNew= set(dataNew.pop('tags'))
+      self.cursor.execute(f"SELECT tag FROM tags WHERE id == '{docID}'")
+      tagsOld= {i[0] for i in self.cursor.fetchall()}
+      if tagsOld.difference(tagsNew):
+        cmd = f"DELETE FROM tags WHERE id == '{docID}' and tag == ?"
+        self.cursor.executemany(cmd, [(i,) for i in tagsOld.difference(tagsNew)])
+        changesDict['tags'] = ','.join(tagsOld)
+      if tagsNew.difference(tagsOld):
+        change = tagsNew.difference(tagsOld)
+        self.cursor.executemany("INSERT INTO tags VALUES (?, ?);", zip([docID]*len(change), change))
+        changesDict['tags'] = ','.join(tagsOld)
+    if 'qrCodes' in dataNew:
+      qrCodesNew= set(dataNew.pop('qrCodes'))
+      self.cursor.execute(f"SELECT qrCode FROM qrCodes WHERE id == '{docID}'")
+      qrCodesOld= {i[0] for i in self.cursor.fetchall()}
+      if qrCodesOld.difference(qrCodesNew):
+        cmd = f"DELETE FROM qrCodes WHERE id == '{docID}' and qrCode == ?"
+        self.cursor.executemany(cmd, [(i,) for i in qrCodesOld.difference(qrCodesNew)])
+        changesDict['qrCodes'] = ','.join(qrCodesOld)
+      if qrCodesNew.difference(qrCodesOld):
+        change = qrCodesNew.difference(qrCodesOld)
+        self.cursor.executemany("INSERT INTO qrCodes VALUES (?, ?);", zip([docID]*len(change), change))
+        changesDict['qrCodes'] = ','.join(qrCodesOld)
     # separate into main and properties
     mainNew    = {key: dataNew.pop(key) for key in MAIN_ORDER if key in dataNew}
     branchNew  = dataNew.pop('branch',{})
